Move _get_external_ip from thorctl.sshconn to thorctl.worker_pool
This seems a slightly more natural place for it, since it can be used for
things other than SSH, like listing instances.
